requirements.txt
streamlit
requests
Pillow
io
json
python-multipart  # Required for form data handling in FastAPI
